fixing bugs

+++ app/src/main/java/com/example/tambahinaja/admin/AdminActivity.kt
@@ -169,7 +169,7 @@ class AdminActivity : AppCompatActivity() {
                     yesButton {
                         val values  = hashMapOf(
                                 "description" to etTournamentDescription.text.toString(),
-                                "price" to etTournamentPrice.text.toString(),
+                                "price" to etTournamentPrice.text.toString().toLong(),
                                 "reward1" to etTournamentReward1.text.toString(),
                                 "reward2" to etTournamentReward2.text.toString(),
                                 "reward3" to etTournamentReward3.text.toString(),
